third_party/WebKit/LayoutTests/fast/events/popup-blocked-from-fake-button-click-expected.txt

PASS win is null
PASS successfullyParsed is true

TEST COMPLETE
